refactor(ui): consolidate stores, normalize file & fn naming
