+++ ui/src/components/FollowersList.vue
@@ -34,7 +34,7 @@
 
 <script setup lang="ts">
 import { followers } from "@/services/mewsfeed-dna";
-import { useProfilesStore } from "@/services/profiles-store";
+import { useProfilesStore } from "@/stores/profiles";
 import { PROFILE_FIELDS } from "@/types/types";
 import { showError } from "@/utils/notification";
 import { AgentPubKey } from "@holochain/client";
