+++ ui/src/components/ProfilePopup.vue
@@ -59,7 +59,7 @@
 </template>
 
 <script setup lang="ts">
-import { useProfilesStore } from "@/services/profiles-store";
+import { useProfilesStore } from "@/stores/profiles";
 import { PROFILE_FIELDS } from "@/types/types";
 import { isSameHash } from "@/utils/hash";
 import { showError } from "@/utils/notification";
