+++ ui/src/pages/MewsFeed.vue
@@ -19,15 +19,15 @@
 <script setup lang="ts">
 import CreateMewField from "@/components/CreateMewField.vue";
 import MewList from "@/components/MewList.vue";
-import { useMewsfeedStore } from "@/stores";
+import { useMewsfeedStore } from "@/stores/mewsfeed";
 import { pageHeightCorrection } from "@/utils/page-layout";
 import { ActionHash } from "@holochain/client";
 import { onMounted } from "vue";
 import { MewTypeName } from "@/types/types";
 
-const store = useMewsfeedStore();
-onMounted(store.fetchMewsFeed);
+const mewsfeedStore = useMewsfeedStore();
+onMounted(mewsfeedStore.fetchMewsFeed);
 
-const onToggleLickMew = (hash: ActionHash) => store.reloadMew(hash);
-const onPublishMew = async () => store.fetchMewsFeed();
+const onToggleLickMew = (hash: ActionHash) => mewsfeedStore.reloadMew(hash);
+const onPublishMew = async () => mewsfeedStore.fetchMewsFeed();
 </script>
